Update PlayerDto namespace and add explicit operator
I was going to use AutoMapper but I came across a vid by CodeWrinkles explaining performance issues and providing alternatives. I really liked the implicit/explicit operators and I'll be using them rn.
