+++ Interfaces/Services/Players/IPlayerService.cs
@@ -1,4 +1,4 @@
-﻿using ShapeDungeon.DTOs.Player;
+﻿using ShapeDungeon.DTOs.Players;
 
 namespace ShapeDungeon.Interfaces.Services.Players
 {
@@ -6,6 +6,6 @@ namespace ShapeDungeon.Interfaces.Services.Players
     {
         Task<bool> CreatePlayerAsync(PlayerDto pDto);
         Task<IEnumerable<PlayerDto>> GetAllPlayersAsync();
-        Task<PlayerDto?> GetPlayerAsync(string name);
+        Task<PlayerDto> GetPlayerAsync(string name);
     }
 }
