+++ Repos/PlayerRepository.cs
@@ -27,21 +27,21 @@ namespace ShapeDungeon.Repos
         /// There can be no active players.
         /// </summary>
         /// <returns>The active player or null.</returns>
-        public async Task<IPlayer?> GetActive()
+        public async Task<Player?> GetActive()
             => await this.Context.Players.SingleOrDefaultAsync(x => x.IsActive);
 
         /// <summary>
         /// All players have names.
         /// </summary>
         /// <returns>Player if one exists with given name, otherwise null.</returns>
-        public async Task<IPlayer?> GetByName(string name)
+        public async Task<Player?> GetByName(string name)
             => await this.Context.Players.FirstOrDefaultAsync(x => x.Name == name);
 
         /// <summary>
         /// Many players can be present and selected.
         /// </summary>
         /// <returns>List of all created players.</returns>
-        public async Task<IEnumerable<IPlayer>> GetAll()
+        public async Task<IEnumerable<Player>> GetAll()
             => await this.Context.Players.ToListAsync();
 
         /// <summary>
@@ -52,10 +52,7 @@ namespace ShapeDungeon.Repos
         public async Task<bool> DoesNameExist(string name)
             => await this.Context.Players.AnyAsync(x => x.Name == name);
 
-        public async Task AddAsync(IPlayer player)
-        {
-            var playerToAdd = (Player)player;
-            await this.Context.Players.AddAsync(playerToAdd);
-        }
+        public async Task AddAsync(Player player)
+            => await this.Context.Players.AddAsync(player);
     }
 }
