+++ Views/Home/_Player.cshtml
@@ -1,4 +1,4 @@
-﻿@using ShapeDungeon.DTOs.Player
+﻿@using ShapeDungeon.DTOs.Players
 @using ShapeDungeon.Entities.Enums
 
 @model PlayerDto
