+++ Views/Player/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@using ShapeDungeon.DTOs.Player;
+﻿@using ShapeDungeon.DTOs.Players;
 
 @model PlayerDto
